Got Nanny Page Working
Click add note and the task will toggle open and close.
Added a text area for the note.
Click the checkmark and it appends the h2 elements.
This needs refinement as it appends every tasks h2.
I think i need to use the data attribute, but not sure how to do that.

Closes #70
Closes #71